Fixes #24037 - Refactor EntitlementsInlineEditFormatter
