+++ webpack/move_to_foreman/components/common/table/formatters/__tests__/EntitlementsInlineEditFormatter.test.js
